01-nlp-synthetic-languages/master-data-science-portfolio/02-real-time-fraud-detection/dashboard/requirements.txt
streamlit
kafka-python==2.2.14